requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
fastapi==0.115.12
filelock==3.18.0
fsspec==2025.5.1
h11==0.16.0
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
opencv-python==4.11.0.86
pillow==11.2.1
pydantic==2.11.5
pydantic_core==2.33.2
python-multipart==0.0.20
requests==2.32.3
sniffio==1.3.1
starlette==0.46.2
sympy==1.14.0
torch==2.7.0
torchaudio==2.7.0
torchvision==0.22.0
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.3